remove unnecessary internal header file
Without any shared code between the primary library and a VFS API,
there's no need for an internal header file at the moment, so we
can eliminate it.
