+++ lib/Makefile.am
@@ -24,7 +24,7 @@ AM_LDFLAGS = $(WARN_LDFLAGS)
 
 lib_LTLIBRARIES = libprojfs.la
 
-libprojfs_la_SOURCES = projfs.c projfs_i.h \
+libprojfs_la_SOURCES = projfs.c \
 		       $(top_srcdir)/include/projfs.h \
 		       $(top_srcdir)/include/projfs_notify.h
